refactor: replace inline file scanning logic with shared utility function for improved code reuse and maintainability in drag-and-drop and upload components
